+++ src/db/ConstraintFinderTrait.php
@@ -31,54 +31,69 @@ trait ConstraintFinderTrait
 {
     /**
      * Returns the metadata of the given type for the given table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param MetadataType $type metadata type.
      * @param bool $refresh whether to reload the table metadata even if it is found in the cache.
+     *
      * @return mixed metadata.
      */
     abstract protected function getTableMetadata(string $name, MetadataType $type, bool $refresh);
 
     /**
      * Returns the metadata of the given type for all tables in the given schema.
-     * @param string $schema the schema of the metadata. Defaults to empty string, meaning the current or default schema name.
+     *
+     * @param string $schema the schema of the metadata. Defaults to empty string, meaning the current or default schema
+     * name.
      * @param MetadataType $type metadata type.
-     * @param bool $refresh whether to fetch the latest available table metadata. If this is `false`,
-     * cached data may be returned if available.
+     * @param bool $refresh whether to fetch the latest available table metadata. If this is `false`, cached data may be
+     * returned if available.
+     *
      * @return array array of metadata.
      */
     abstract protected function getSchemaMetadata(string $schema, MetadataType $type, bool $refresh);
 
     /**
      * Loads a primary key for the given table.
+     *
      * @param string $tableName table name.
+     *
      * @return Constraint|null primary key for the given table, `null` if the table has no primary key.
      */
     abstract protected function loadTablePrimaryKey($tableName);
 
     /**
      * Loads all foreign keys for the given table.
+     *
      * @param string $tableName table name.
+     *
      * @return ForeignKeyConstraint[] foreign keys for the given table.
      */
     abstract protected function loadTableForeignKeys($tableName);
 
     /**
      * Loads all indexes for the given table.
+     *
      * @param string $tableName table name.
+     *
      * @return IndexConstraint[] indexes for the given table.
      */
     abstract protected function loadTableIndexes($tableName);
 
     /**
      * Loads all unique constraints for the given table.
+     *
      * @param string $tableName table name.
+     *
      * @return Constraint[] unique constraints for the given table.
      */
     abstract protected function loadTableUniques($tableName);
 
     /**
      * Loads all check constraints for the given table.
+     *
      * @param string $tableName table name.
+     *
      * @return CheckConstraint[] check constraints for the given table.
      */
     abstract protected function loadTableChecks($tableName);
@@ -87,151 +102,182 @@ trait ConstraintFinderTrait
      * Loads all default value constraints for the given table.
      *
      * @param string $tableName table name.
+     *
      * @return DefaultValueConstraint[] default value constraints for the given table.
      */
     abstract protected function loadTableDefaultValues($tableName);
 
     /**
      * Obtains the primary key for the named table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param bool $refresh whether to reload the information even if it is found in the cache.
+     *
      * @return Constraint|null table primary key, `null` if the table has no primary key.
      */
     public function getTablePrimaryKey($name, $refresh = false)
     {
-        return $this->getTableMetadata($name, MetadataType::PrimaryKey, $refresh);
+        return $this->getTableMetadata($name, MetadataType::PRIMARY_KEY, $refresh);
     }
 
     /**
      * Returns primary keys for all tables in the database.
-     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema name.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema
+     * name.
      * @param bool $refresh whether to fetch the latest available table schemas. If this is `false`,
      * cached data may be returned if available.
+     *
      * @return Constraint[] primary keys for all tables in the database.
      * Each array element is an instance of [[Constraint]] or its child class.
      */
     public function getSchemaPrimaryKeys($schema = '', $refresh = false)
     {
-        return $this->getSchemaMetadata($schema, MetadataType::PrimaryKey, $refresh);
+        return $this->getSchemaMetadata($schema, MetadataType::PRIMARY_KEY, $refresh);
     }
 
     /**
      * Obtains the foreign keys information for the named table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param bool $refresh whether to reload the information even if it is found in the cache.
+     *
      * @return ForeignKeyConstraint[] table foreign keys.
      */
     public function getTableForeignKeys($name, $refresh = false)
     {
-        return $this->getTableMetadata($name, MetadataType::ForeignKeys, $refresh);
+        return $this->getTableMetadata($name, MetadataType::FOREIGN_KEYS, $refresh);
     }
 
     /**
      * Returns foreign keys for all tables in the database.
-     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema name.
-     * @param bool $refresh whether to fetch the latest available table schemas. If this is false,
-     * cached data may be returned if available.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema
+     * name.
+     * @param bool $refresh whether to fetch the latest available table schemas. If this is false, cached data may be
+     * returned if available.
+     *
      * @return ForeignKeyConstraint[][] foreign keys for all tables in the database.
      * Each array element is an array of [[ForeignKeyConstraint]] or its child classes.
      */
     public function getSchemaForeignKeys($schema = '', $refresh = false)
     {
-        return $this->getSchemaMetadata($schema, MetadataType::ForeignKeys, $refresh);
+        return $this->getSchemaMetadata($schema, MetadataType::FOREIGN_KEYS, $refresh);
     }
 
     /**
      * Obtains the indexes information for the named table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param bool $refresh whether to reload the information even if it is found in the cache.
+     *
      * @return IndexConstraint[] table indexes.
      */
     public function getTableIndexes($name, $refresh = false)
     {
-        return $this->getTableMetadata($name, MetadataType::Indexes, $refresh);
+        return $this->getTableMetadata($name, MetadataType::INDEXES, $refresh);
     }
 
     /**
      * Returns indexes for all tables in the database.
-     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema name.
-     * @param bool $refresh whether to fetch the latest available table schemas. If this is false,
-     * cached data may be returned if available.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema
+     * name.
+     * @param bool $refresh whether to fetch the latest available table schemas. If this is false, cached data may be
+     * returned if available.
+     *
      * @return IndexConstraint[][] indexes for all tables in the database.
      * Each array element is an array of [[IndexConstraint]] or its child classes.
      */
     public function getSchemaIndexes($schema = '', $refresh = false)
     {
-        return $this->getSchemaMetadata($schema, MetadataType::Indexes, $refresh);
+        return $this->getSchemaMetadata($schema, MetadataType::INDEXES, $refresh);
     }
 
     /**
      * Obtains the unique constraints information for the named table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param bool $refresh whether to reload the information even if it is found in the cache.
+     *
      * @return Constraint[] table unique constraints.
      */
     public function getTableUniques($name, $refresh = false)
     {
-        return $this->getTableMetadata($name, MetadataType::Uniques, $refresh);
+        return $this->getTableMetadata($name, MetadataType::UNIQUES, $refresh);
     }
 
     /**
      * Returns unique constraints for all tables in the database.
-     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema name.
-     * @param bool $refresh whether to fetch the latest available table schemas. If this is false,
-     * cached data may be returned if available.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema
+     * name.
+     * @param bool $refresh whether to fetch the latest available table schemas. If this is false, cached data may be
+     * returned if available.
+     *
      * @return Constraint[][] unique constraints for all tables in the database.
      * Each array element is an array of [[Constraint]] or its child classes.
      */
     public function getSchemaUniques($schema = '', $refresh = false)
     {
-        return $this->getSchemaMetadata($schema, MetadataType::Uniques, $refresh);
+        return $this->getSchemaMetadata($schema, MetadataType::UNIQUES, $refresh);
     }
 
     /**
      * Obtains the check constraints information for the named table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param bool $refresh whether to reload the information even if it is found in the cache.
+     *
      * @return CheckConstraint[] table check constraints.
      */
     public function getTableChecks($name, $refresh = false)
     {
-        return $this->getTableMetadata($name, MetadataType::Checks, $refresh);
+        return $this->getTableMetadata($name, MetadataType::CHECKS, $refresh);
     }
 
     /**
      * Returns check constraints for all tables in the database.
-     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema name.
-     * @param bool $refresh whether to fetch the latest available table schemas. If this is false,
-     * cached data may be returned if available.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema
+     * name.
+     * @param bool $refresh whether to fetch the latest available table schemas. If this is false, cached data may be
+     * returned if available.
+     *
      * @return CheckConstraint[][] check constraints for all tables in the database.
      * Each array element is an array of [[CheckConstraint]] or its child classes.
      */
     public function getSchemaChecks($schema = '', $refresh = false)
     {
-        return $this->getSchemaMetadata($schema, MetadataType::Checks, $refresh);
+        return $this->getSchemaMetadata($schema, MetadataType::CHECKS, $refresh);
     }
 
     /**
      * Obtains the default value constraints information for the named table.
+     *
      * @param string $name table name. The table name may contain schema name if any. Do not quote the table name.
      * @param bool $refresh whether to reload the information even if it is found in the cache.
+     *
      * @return DefaultValueConstraint[] table default value constraints.
      */
     public function getTableDefaultValues($name, $refresh = false)
     {
-        return $this->getTableMetadata($name, MetadataType::DefaultValues, $refresh);
+        return $this->getTableMetadata($name, MetadataType::DEFAULT_VALUES, $refresh);
     }
 
     /**
      * Returns default value constraints for all tables in the database.
-     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema name.
+     *
+     * @param string $schema the schema of the tables. Defaults to empty string, meaning the current or default schema
+     * name.
      * @param bool $refresh whether to fetch the latest available table schemas. If this is false,
      * cached data may be returned if available.
+     *
      * @return DefaultValueConstraint[] default value constraints for all tables in the database.
      * Each array element is an array of [[DefaultValueConstraint]] or its child classes.
      */
     public function getSchemaDefaultValues($schema = '', $refresh = false)
     {
-        return $this->getSchemaMetadata($schema, MetadataType::DefaultValues, $refresh);
+        return $this->getSchemaMetadata($schema, MetadataType::DEFAULT_VALUES, $refresh);
     }
 }