+++ src/db/pgsql/Schema.php
@@ -25,7 +25,6 @@ use yii\db\ViewFinderTrait;
 use yii\helpers\ArrayHelper;
 
 use function array_change_key_case;
-use function array_merge;
 use function explode;
 use function implode;
 use function in_array;
@@ -236,7 +235,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
      */
     protected function loadTablePrimaryKey($tableName)
     {
-        return $this->loadTableConstraints($tableName, MetadataType::PrimaryKey);
+        return $this->loadTableConstraints($tableName, MetadataType::PRIMARY_KEY);
     }
 
     /**
@@ -244,7 +243,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
      */
     protected function loadTableForeignKeys($tableName)
     {
-        return $this->loadTableConstraints($tableName, MetadataType::ForeignKeys);
+        return $this->loadTableConstraints($tableName, MetadataType::FOREIGN_KEYS);
     }
 
     /**
@@ -309,7 +308,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
      */
     protected function loadTableUniques($tableName)
     {
-        return $this->loadTableConstraints($tableName, MetadataType::Uniques);
+        return $this->loadTableConstraints($tableName, MetadataType::UNIQUES);
     }
 
     /**
@@ -317,7 +316,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
      */
     protected function loadTableChecks($tableName)
     {
-        return $this->loadTableConstraints($tableName, MetadataType::Checks);
+        return $this->loadTableConstraints($tableName, MetadataType::CHECKS);
     }
 
     /**
@@ -441,7 +440,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
         }
 
         foreach ($constraints as $name => $constraint) {
-            $table->foreignKeys[$name] = array_merge([$constraint['tableName']], $constraint['columns']);
+            $table->foreignKeys[$name] = [$constraint['tableName'], ...$constraint['columns']];
         }
     }
 
@@ -818,17 +817,17 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
         $constraints = ArrayHelper::index($constraints, null, ['type', 'name']);
 
         $result = [
-            MetadataType::PrimaryKey->value => null,
-            MetadataType::ForeignKeys->value => [],
-            MetadataType::Uniques->value => [],
-            MetadataType::Checks->value => [],
+            MetadataType::CHECKS->value => [],
+            MetadataType::FOREIGN_KEYS->value => [],
+            MetadataType::PRIMARY_KEY->value => null,
+            MetadataType::UNIQUES->value => [],
         ];
 
         foreach ($constraints as $type => $names) {
             foreach ($names as $name => $constraint) {
                 switch ($type) {
                     case 'p':
-                        $result[MetadataType::PrimaryKey->value] = new Constraint(
+                        $result[MetadataType::PRIMARY_KEY->value] = new Constraint(
                             [
                                 'name' => $name,
                                 'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
@@ -836,7 +835,7 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
                         );
                         break;
                     case 'f':
-                        $result[MetadataType::ForeignKeys->value][] = new ForeignKeyConstraint(
+                        $result[MetadataType::FOREIGN_KEYS->value][] = new ForeignKeyConstraint(
                             [
                                 'name' => $name,
                                 'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
@@ -849,13 +848,13 @@ class Schema extends \yii\db\Schema implements ConstraintFinderInterface
                         );
                         break;
                     case 'u':
-                        $result[MetadataType::Uniques->value][] = new Constraint([
+                        $result[MetadataType::UNIQUES->value][] = new Constraint([
                             'name' => $name,
                             'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
                         ]);
                         break;
                     case 'c':
-                        $result[MetadataType::Checks->value][] = new CheckConstraint([
+                        $result[MetadataType::CHECKS->value][] = new CheckConstraint([
                             'name' => $name,
                             'columnNames' => ArrayHelper::getColumn($constraint, 'column_name'),
                             'expression' => $constraint[0]['check_expr'],